script that implements the Macenko method
