+++ bin/norm_Macenko.py
@@ -13,8 +13,7 @@ Use with python via e.g https://anaconda.org/conda-forge/python-spams
 from __future__ import division
 
 import numpy as np
-import stain_utils as ut
-
+import spams
 
 def get_stain_matrix(I, beta=0.15, alpha=1):
     """
@@ -24,7 +23,7 @@ def get_stain_matrix(I, beta=0.15, alpha=1):
     :param alpha:
     :return:
     """
-    OD = ut.RGB_to_OD(I).reshape((-1, 3))
+    OD = RGB_to_OD(I).reshape((-1, 3))
     OD = (OD[(OD > beta).any(axis=1), :])
     _, V = np.linalg.eigh(np.cov(OD, rowvar=False))
     V = V[:, [2, 1]]
@@ -40,8 +39,48 @@ def get_stain_matrix(I, beta=0.15, alpha=1):
         HE = np.array([v1, v2])
     else:
         HE = np.array([v2, v1])
-    return ut.normalize_rows(HE)
+    return normalize_rows(HE)
 
+def standardize_brightness(I):
+    """
+    :param I:
+    :return:
+    """
+    p = np.percentile(I, 90)
+    return np.clip(I * 255.0 / p, 0, 255).astype(np.uint8)
+
+def RGB_to_OD(I):
+    """
+    Convert from RGB to optical density
+    :param I:
+    :return:
+    """
+    mask = (I == 0)
+    I[mask] = 1
+    return -1 * np.log(I / 255)
+
+def normalize_rows(A):
+    """
+    Normalize rows of an array
+    :param A:
+    :return:
+    """
+    return A / np.linalg.norm(A, axis=1)[:, None]
+
+def get_concentrations(I, stain_matrix, lamda=0.01):
+    """
+    Get concentrations, a npix x 2 matrix
+    :param I:
+    :param stain_matrix: a 2x3 stain matrix
+    :return:
+    """
+    OD = RGB_to_OD(I).reshape((-1, 3))
+    try:
+        #limit Lasso threads to 1 to avoid thread inception (default is -1, takes all cores)
+        temp = spams.lasso(OD.T, D=stain_matrix.T, mode=2, lambda1=lamda, pos=True, numThreads=1).toarray().T
+    except:
+       temp = 0
+    return temp
 
 ###
 
@@ -55,28 +94,26 @@ class Normalizer(object):
         self.target_concentrations = None
 
     def fit(self, target):
-        target = ut.standardize_brightness(target)
+        """
+        Fit to a target image
+        :param target: Image RGB uint8
+        :return:
+        """
+        target = standardize_brightness(target)
         self.stain_matrix_target = get_stain_matrix(target)
-        self.target_concentrations = ut.get_concentrations(target, self.stain_matrix_target)
-
-    def target_stains(self):
-        return ut.OD_to_RGB(self.stain_matrix_target)
+        self.target_concentrations = get_concentrations(target, self.stain_matrix_target)
 
     def transform(self, I):
-        I = ut.standardize_brightness(I)
+        """
+        Transform of the image
+        :param I: Image RGB uint8
+        :return:
+        """
+        I = standardize_brightness(I)
         stain_matrix_source = get_stain_matrix(I)
-        source_concentrations = ut.get_concentrations(I, stain_matrix_source)
+        source_concentrations = get_concentrations(I, stain_matrix_source)
         maxC_source = np.percentile(source_concentrations, 99, axis=0).reshape((1, 2))
         maxC_target = np.percentile(self.target_concentrations, 99, axis=0).reshape((1, 2))
         source_concentrations *= (maxC_target / maxC_source)
         return (255 * np.exp(-1 * np.dot(source_concentrations, self.stain_matrix_target).reshape(I.shape))).astype(
             np.uint8)
-
-    def hematoxylin(self, I):
-        I = ut.standardize_brightness(I)
-        h, w, c = I.shape
-        stain_matrix_source = get_stain_matrix(I)
-        source_concentrations = ut.get_concentrations(I, stain_matrix_source)
-        H = source_concentrations[:, 0].reshape(h, w)
-        H = np.exp(-1 * H)
-        return H
